requirements.txt
flask
flask-cors
pdfplumber
selenium
webdriver-manager
requests
beautifulsoup4
python-dotenv
redis
groq